requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dominate==2.9.1
filelock==3.13.1
fonttools==4.60.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
pillow==11.0.0
platformdirs==4.4.0
polars==1.33.1
protobuf==6.32.1
psutil==7.1.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.2
sentry-sdk==2.38.0
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tifffile==2025.9.20
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128
typing-inspection==0.4.1
typing_extensions==4.12.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
wandb==0.22.0
